requirements.txt
aiofiles==23.1.0
fastapi==0.96.0
uvicorn==0.22.0
sqlalchemy-2.0.15